Set status to error if notification fails to send
